update db to 7.15

+++ sql/drop/h2_engine_7.15.0.sql
@@ -22,8 +22,8 @@ drop index ACT_IDX_DEC_DEF_TENANT_ID;
 drop index ACT_IDX_DEC_DEF_REQ_ID;
 drop index ACT_IDX_DEC_REQ_DEF_TENANT_ID;
 
-drop table ACT_RE_DECISION_DEF;
-drop table ACT_RE_DECISION_REQ_DEF;
+drop table ACT_RE_DECISION_DEF if exists;
+drop table ACT_RE_DECISION_REQ_DEF if exists;
 --
 -- Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
 -- under one or more contributor license agreements. See the NOTICE file
@@ -45,46 +45,36 @@ drop index ACT_IDX_CASE_EXEC_BUSKEY;
 drop index ACT_IDX_CASE_DEF_TENANT_ID;
 drop index ACT_IDX_CASE_EXEC_TENANT_ID;
 
-drop index ACT_IDX_CASE_EXE_CASE_INST;
-drop index ACT_IDX_CASE_EXE_PARENT;
-drop index ACT_IDX_CASE_EXE_CASE_DEF;
-drop index ACT_IDX_VAR_CASE_EXE;
-drop index ACT_IDX_VAR_CASE_INST_ID;
-drop index ACT_IDX_TASK_CASE_EXEC;
-drop index ACT_IDX_TASK_CASE_DEF_ID;
-drop index ACT_IDX_CASE_SENTRY_CASE_INST;
-drop index ACT_IDX_CASE_SENTRY_CASE_EXEC;
+alter table ACT_RU_CASE_EXECUTION
+    drop constraint ACT_FK_CASE_EXE_CASE_INST;
 
 alter table ACT_RU_CASE_EXECUTION
-    drop CONSTRAINT ACT_FK_CASE_EXE_CASE_INST;
+    drop constraint ACT_FK_CASE_EXE_PARENT;
 
 alter table ACT_RU_CASE_EXECUTION
-    drop CONSTRAINT ACT_FK_CASE_EXE_PARENT;
-
-alter table ACT_RU_CASE_EXECUTION
-    drop CONSTRAINT ACT_FK_CASE_EXE_CASE_DEF;
+    drop constraint ACT_FK_CASE_EXE_CASE_DEF;
 
 alter table ACT_RU_VARIABLE
-    drop CONSTRAINT ACT_FK_VAR_CASE_EXE;
+    drop constraint ACT_FK_VAR_CASE_EXE;
 
 alter table ACT_RU_VARIABLE
-    drop CONSTRAINT ACT_FK_VAR_CASE_INST;
+    drop constraint ACT_FK_VAR_CASE_INST;
 
 alter table ACT_RU_TASK
-    drop CONSTRAINT ACT_FK_TASK_CASE_EXE;
+    drop constraint ACT_FK_TASK_CASE_EXE;
 
 alter table ACT_RU_TASK
-    drop CONSTRAINT ACT_FK_TASK_CASE_DEF;
+    drop constraint ACT_FK_TASK_CASE_DEF;
 
 alter table ACT_RU_CASE_SENTRY_PART
-    drop CONSTRAINT ACT_FK_CASE_SENTRY_CASE_INST;
+    drop constraint ACT_FK_CASE_SENTRY_CASE_INST;
 
 alter table ACT_RU_CASE_SENTRY_PART
-    drop CONSTRAINT ACT_FK_CASE_SENTRY_CASE_EXEC;
+    drop constraint ACT_FK_CASE_SENTRY_CASE_EXEC;
 
-drop table ACT_RE_CASE_DEF;
-drop table ACT_RU_CASE_EXECUTION;
-drop table ACT_RU_CASE_SENTRY_PART;
+drop table ACT_RE_CASE_DEF if exists;
+drop table ACT_RU_CASE_EXECUTION if exists;
+drop table ACT_RU_CASE_SENTRY_PART if exists;
 --
 -- Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
 -- under one or more contributor license agreements. See the NOTICE file
@@ -102,26 +92,9 @@ drop table ACT_RU_CASE_SENTRY_PART;
 -- limitations under the License.
 --
 
-drop index ACT_IDX_BYTEARRAY_RM_TIME;
 drop index ACT_IDX_BYTEARRAY_ROOT_PI;
-drop index ACT_IDX_BYTEAR_DEPL ;
-drop index ACT_IDX_EXE_ROOT_PI;
-drop index ACT_IDX_EXE_PROCINST ;
-drop index ACT_IDX_EXE_PARENT ;
-drop index ACT_IDX_EXE_SUPER;
-drop index ACT_IDX_EXE_PROCDEF;
-drop index ACT_IDX_TSKASS_TASK;
-drop index ACT_IDX_TASK_EXEC;
-drop index ACT_IDX_TASK_PROCINST;
-drop index ACT_IDX_TASK_PROCDEF;
-drop index ACT_IDX_VAR_EXE;
-drop index ACT_IDX_VAR_PROCINST;
-drop index ACT_IDX_VAR_BYTEARRAY;
-drop index ACT_IDX_JOB_EXCEPTION;
-drop index ACT_IDX_JOB_PROCINST;
-drop index ACT_IDX_INC_CONFIGURATION;
-drop index ACT_IDX_AUTH_GROUP_ID;
-
+drop index ACT_IDX_BYTEARRAY_RM_TIME;
+drop index ACT_IDX_EXEC_ROOT_PI;
 drop index ACT_IDX_EXEC_BUSKEY;
 drop index ACT_IDX_TASK_CREATE;
 drop index ACT_IDX_TASK_ASSIGNEE;
@@ -130,20 +103,9 @@ drop index ACT_IDX_IDENT_LNK_USER;
 drop index ACT_IDX_IDENT_LNK_GROUP;
 drop index ACT_IDX_VARIABLE_TASK_ID;
 drop index ACT_IDX_VARIABLE_TASK_NAME_TYPE;
-
--- new metric milliseconds column
-DROP INDEX ACT_IDX_METER_LOG_MS;
-DROP INDEX ACT_IDX_METER_LOG_NAME_MS;
-DROP INDEX ACT_IDX_METER_LOG_REPORT;
-
--- old metric timestamp column
-DROP INDEX ACT_IDX_METER_LOG_TIME;
-DROP INDEX ACT_IDX_METER_LOG;
-
-drop index ACT_IDX_EXT_TASK_TOPIC;
-
-drop index ACT_IDX_JOB_EXECUTION_ID;
-drop index ACT_IDX_JOB_HANDLER;
+drop index ACT_IDX_INC_CONFIGURATION;
+drop index ACT_IDX_JOB_PROCINST;
+drop index ACT_IDX_AUTH_GROUP_ID;
 
 alter table ACT_GE_BYTEARRAY
     drop constraint ACT_FK_BYTEARR_DEPL;
@@ -236,7 +198,6 @@ alter table ACT_RU_VARIABLE
     drop CONSTRAINT ACT_FK_VAR_BATCH;
 
 drop index ACT_IDX_EVENT_SUBSCR_CONFIG_;
-drop index ACT_IDX_EVENT_SUBSCR;
 drop index ACT_IDX_ATHRZ_PROCEDEF;
 
 -- indexes for deadlock problems - https://app.camunda.com/jira/browse/CAM-2567
@@ -246,7 +207,21 @@ drop index ACT_IDX_INC_PROCDEFID;
 drop index ACT_IDX_INC_PROCINSTID;
 drop index ACT_IDX_INC_ROOTCAUSEINCID;
 drop index ACT_IDX_INC_JOB_DEF;
+
+-- new metric milliseconds column
+DROP INDEX ACT_IDX_METER_LOG_MS;
+DROP INDEX ACT_IDX_METER_LOG_NAME_MS;
+DROP INDEX ACT_IDX_METER_LOG_REPORT;
+
+-- old metric timestamp column
+DROP INDEX ACT_IDX_METER_LOG_TIME;
+DROP INDEX ACT_IDX_METER_LOG;
+
+-- task metric timestamp column
+drop index ACT_IDX_TASK_METER_LOG_TIME;
+
 drop index ACT_IDX_AUTH_RESOURCE_ID;
+drop index ACT_IDX_EXT_TASK_TOPIC;
 drop index ACT_IDX_EXT_TASK_EXEC;
 
 drop index ACT_IDX_BYTEARRAY_NAME;
@@ -276,29 +251,34 @@ drop index ACT_IDX_BATCH_JOB_DEF;
 
 drop index ACT_IDX_PROCDEF_VER_TAG;
 
+drop index ACT_IDX_JOB_EXECUTION_ID;
+drop index ACT_IDX_JOB_HANDLER;
+
 drop index ACT_IDX_AUTH_ROOT_PI;
 drop index ACT_IDX_AUTH_RM_TIME;
 
 drop index ACT_IDX_BATCH_ID;
 
-drop table ACT_GE_PROPERTY;
-drop table ACT_GE_BYTEARRAY;
-drop table ACT_RE_DEPLOYMENT;
-drop table ACT_RE_PROCDEF;
-drop table ACT_RU_EXECUTION;
-drop table ACT_RU_JOB;
-drop table ACT_RU_JOBDEF;
-drop table ACT_RU_TASK;
-drop table ACT_RU_IDENTITYLINK;
-drop table ACT_RU_VARIABLE;
-drop table ACT_RU_EVENT_SUBSCR;
-drop table ACT_RU_INCIDENT;
-drop table ACT_RU_AUTHORIZATION;
-drop table ACT_RU_FILTER;
-drop table ACT_RU_METER_LOG;
-drop table ACT_RU_EXT_TASK;
-drop table ACT_RU_BATCH;
-drop table ACT_GE_SCHEMA_LOG;
+drop table ACT_GE_PROPERTY if exists;
+drop table ACT_GE_BYTEARRAY if exists;
+drop table ACT_RE_DEPLOYMENT if exists;
+drop table ACT_RU_EXECUTION if exists;
+drop table ACT_RU_JOB if exists;
+drop table ACT_RU_JOBDEF if exists;
+drop table ACT_RE_PROCDEF if exists;
+drop table ACT_RU_TASK if exists;
+drop table ACT_RU_IDENTITYLINK if exists;
+drop table ACT_RU_VARIABLE if exists;
+drop table ACT_RU_EVENT_SUBSCR if exists;
+drop table ACT_RU_INCIDENT if exists;
+drop table ACT_RU_AUTHORIZATION if exists;
+drop table ACT_RU_FILTER if exists;
+drop table ACT_RU_METER_LOG if exists;
+drop table ACT_RU_TASK_METER_LOG if exists;
+drop table ACT_RU_EXT_TASK if exists;
+drop table ACT_RU_BATCH if exists;
+drop table ACT_GE_SCHEMA_LOG if exists;
+
 --
 -- Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
 -- under one or more contributor license agreements. See the NOTICE file
@@ -340,11 +320,11 @@ drop index ACT_IDX_HI_DEC_OUT_RULE;
 drop index ACT_IDX_HI_DEC_OUT_ROOT_PI;
 drop index ACT_IDX_HI_DEC_OUT_RM_TIME;
 
-drop table ACT_HI_DECINST;
+drop table ACT_HI_DECINST if exists;
 
-drop table ACT_HI_DEC_IN;
+drop table ACT_HI_DEC_IN if exists;
 
-drop table ACT_HI_DEC_OUT;
+drop table ACT_HI_DEC_OUT if exists;
 --
 -- Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
 -- under one or more contributor license agreements. See the NOTICE file
@@ -370,8 +350,8 @@ drop index ACT_IDX_HI_CAS_A_I_END;
 drop index ACT_IDX_HI_CAS_A_I_COMP;
 drop index ACT_IDX_HI_CAS_A_I_TENANT_ID;
 
-drop table ACT_HI_CASEINST;
-drop table ACT_HI_CASEACTINST;
+drop table ACT_HI_CASEINST if exists;
+drop table ACT_HI_CASEACTINST if exists;
 --
 -- Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
 -- under one or more contributor license agreements. See the NOTICE file
@@ -501,16 +481,16 @@ drop index ACT_IDX_HI_COMMENT_ROOT_PI;
 drop index ACT_IDX_HI_COMMENT_PROCINST;
 drop index ACT_IDX_HI_COMMENT_RM_TIME;
 
-drop table ACT_HI_PROCINST;
-drop table ACT_HI_ACTINST;
-drop table ACT_HI_VARINST;
-drop table ACT_HI_TASKINST;
-drop table ACT_HI_DETAIL;
-drop table ACT_HI_COMMENT;
-drop table ACT_HI_ATTACHMENT;
-drop table ACT_HI_OP_LOG;
-drop table ACT_HI_INCIDENT;
-drop table ACT_HI_JOB_LOG;
-drop table ACT_HI_BATCH;
-drop table ACT_HI_IDENTITYLINK;
-drop table ACT_HI_EXT_TASK_LOG;
+drop table ACT_HI_PROCINST if exists;
+drop table ACT_HI_ACTINST if exists;
+drop table ACT_HI_VARINST if exists;
+drop table ACT_HI_TASKINST if exists;
+drop table ACT_HI_DETAIL if exists;
+drop table ACT_HI_COMMENT if exists;
+drop table ACT_HI_ATTACHMENT if exists;
+drop table ACT_HI_OP_LOG if exists;
+drop table ACT_HI_INCIDENT if exists;
+drop table ACT_HI_JOB_LOG if exists;
+drop table ACT_HI_BATCH if exists;
+drop table ACT_HI_IDENTITYLINK if exists;
+drop table ACT_HI_EXT_TASK_LOG if exists;
